+++ src/main/java/me/melontini/andromeda/entity/vehicle/boats/BoatEntityWithBlock.java
@@ -1,9 +1,7 @@
 package me.melontini.andromeda.entity.vehicle.boats;
 
-import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.vehicle.BoatEntity;
-import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
 
 public class BoatEntityWithBlock extends BoatEntity {
@@ -14,16 +12,8 @@ public class BoatEntityWithBlock extends BoatEntity {
     }
 
     @Override
-    public void updatePassengerPosition(Entity passenger, PositionUpdater positionUpdater) {
-        if (this.hasPassenger(passenger)) {
-            double g = ((this.isRemoved() ? 0.01F : this.getMountedHeightOffset()) + passenger.getHeightOffset());
-
-            Vec3d vec3d = new Vec3d(0.3f, 0.0, 0.0).rotateY(-this.getYaw() * PIby180 - PIby2);
-            passenger.setPosition(this.getX() + vec3d.x, this.getY() + g, this.getZ() + vec3d.z);
-            passenger.setYaw(passenger.getYaw() + this.yawVelocity);
-            passenger.setHeadYaw(passenger.getHeadYaw() + this.yawVelocity);
-            this.copyEntityData(passenger);
-        }
+    protected float getPassengerHorizontalOffset() {
+        return 0.15f;
     }
 
     @Override