+++ src/main/java/me/melontini/andromeda/util/annotations/config/Environment.java
@@ -1,0 +1,7 @@
+package me.melontini.andromeda.util.annotations.config;
+
+public enum Environment {
+    CLIENT,
+    SERVER,
+    BOTH
+}