+++ CHANGELOG.md
@@ -3,10 +3,18 @@
 No hotfixes today. Just lukewarm fixes.
 
 * Fixed the `.andromeda` folder not getting created in some cases.
+* Fixed the item group causing a crash if you enabled an unregistered item and didn't restart Minecraft.
+* Fixed furnace boats going underwater.
+* Changed effect list for "Campfire Effects". You need to specify effects again.
+* GUI particles will no longer flash for a frame after closing their screen.
+* Most features now... feature an environment marker. (Client/Server/Both)
 * Made all options require restart if Compat Mode is on.
 * Fixed some classes appearing as `[]` in the log.
 * Optional data will only get uploaded after mod update or first install.
 * Replaced accessors with AWs.
 * Added success logs to more registries.
 * Improved "Item Frame Tooltips" logic.
+* Switched to builders in AdvancementGeneration.
 * Updated Chinese translation. Courtesy of [Rad233](https://github.com/Rad233).
+* Updated Dark Matter.
+* Some other things.